refactor: separation module
